Move input into client
